refactor: extract local agent dispatch helpers
